requirements.txt
discord.py
imdbpy
dataclasses
asyncio